moved the beacon service to its own module
