+++ frontend/src/app/beacon/beacon.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import {BeaconService} from './beacon.service';
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule
+  ],
+  providers: [BeaconService],
+})
+export class BeaconModule { }
