+++ frontend/src/app/beacon/beacon.service.ts
