+++ frontend/src/app/wallet/wallet.module.ts
@@ -4,13 +4,13 @@ import { CommonModule } from '@angular/common';
 import { WalletRoutingModule } from './wallet-routing.module';
 import { WalletComponent } from './wallet.component';
 import {MatButtonModule} from '@angular/material/button';
-import {BeaconService} from './beacon.service';
 import {MatCardModule} from '@angular/material/card';
 import {FontAwesomeModule} from '@fortawesome/angular-fontawesome';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {ReactiveFormsModule} from '@angular/forms';
 import {MatInputModule} from '@angular/material/input';
 import {TranslateModule} from '@ngx-translate/core';
+import {BeaconModule} from '../beacon/beacon.module';
 
 
 @NgModule({
@@ -27,7 +27,7 @@ import {TranslateModule} from '@ngx-translate/core';
     ReactiveFormsModule,
     MatInputModule,
     TranslateModule,
+    BeaconModule,
   ],
-  providers: [BeaconService],
 })
 export class WalletModule { }
